修改枚举属性为Byte

+++ dal/src/main/java/com/timevale/forward/dal/entity/FileDO.java
@@ -20,7 +20,7 @@ public class FileDO extends BaseDO {
     /**
      * 附件所属类型
      */
-    Integer type;
+    Byte type;
 
     /**
      * 附件名称
